requirements.txt
django
gunicorn
django-heroku
Flask
ptvsd
pandas
plotly
numpy
sympy
dash
dash-bootstrap-components
mpmath
dash_daq
seaborne
matplotlib